// client/src/excel/constants.ts
export enum ETableType {
  OUTDOOR_AREA = 'venkovní plocha',
  FOLIO = 'fóliovník',
  GREENHOUSE = 'skleník',
  HOTBED = 'pařeniště',
  STOCK = 'sklad',
  PICKLING_STATION = 'mořicí stanice',
}

export enum ESummaryTabType {
  Center,
  Wood,
  Table,
}

export enum EModule {
  Nursery = 'NURSERY',
  Seller = 'SELLER'
}